src/requirements.txt
configparser==3.5.0
diff-match-patch==20121119
Django==1.10.7
django-bootstrap3==9.0.0
django-excel==0.0.9
django-import-export==0.5.1
django-taggit==0.22.1
django-tags-input==4.0.0
et-xmlfile==1.0.1
jdcal==1.3
lml==0.0.1
mysqlclient==1.3.10
odfpy==1.3.5
openpyxl==2.4.8
pyexcel==0.5.6
pyexcel-io==0.5.3
pyexcel-webio==0.1.4
pyexcel-xls==0.5.4
pyexcel-xlsx==0.5.4
PyMySQL==0.7.11
PyYAML==3.12
six==1.11.0
tablib==0.12.1
texttable==1.1.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
